Update de.yml
The German word "übergeordnete Aufgabe" is correct but very long and takes a lot of space as a column in the project overview. That's why I suggest "Hauptaufgabe" which is shorter and so you have more space for other columns in the overview. Maybe someone knows an even shorter word as the column will only show the ID.